refactor : 서버폴더 코드 리팩토링
