+++ server/Connect/web3.js
@@ -1,13 +1,10 @@
 const Contract = require('web3-eth-contract');
-const abi =require("./abi.json");
-
+const abi =require("../../client/src/abi/abi.json");
 require("dotenv").config();
-
 try{
     const nft_contract=new Contract(abi,process.env.ADDRESS);
     // nft_contract.setProvider('http://127.0.0.1:7545');
     nft_contract.setProvider('https://goerli.infura.io/v3/b03f802e554f441786b51c437837bfe4');
-    
     module.exports={
         nft_contract
     }
